+++ src/types/button.d.ts
@@ -1,0 +1,10 @@
+export type FloatingButtonProps = {
+  icon: JSX.Element;
+  onClick: () => void;
+};
+
+export type CreateModalButtonProps = {
+  text: string;
+  isDisabled?: boolean;
+  clickButtonHandler: () => void;
+};
