--- src/types/floatingButton.d.ts
@@ -1,4 +1,0 @@
-export type FloatingButtonProps = {
-  icon: JSX.Element;
-  onClick: () => void;
-};
